# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.test/api"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "7208c29037df2b8bb02499ba39ed8a78",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Agregar nuevo usuario",
                "operationId": "262eb5ca1fdea354023c0a8870442c1e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "default": "Login successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "default": "20d338931e8d6bd9466edeba78ea7dce7c7bc01aa5cc5b4735691c50a2fe3228"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener especificamente un usuario",
                "operationId": "6946bbab7898f59a09e3c8e529a38907",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pasamos el ID del usuario a buscar como parámetro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "ID #1",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "operationId": "05028d02ea3e9a99abd6260cfe05b179",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            },
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "name": {
                        "type": "string",
                        "default": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "default": "jdoe@email.com"
                    },
                    "password": {
                        "type": "string",
                        "default": "123456"
                    }
                },
                "type": "object"
            }
        }
    }
}